feat: query/2 uses match if no guards are needed
For queries without guards, we can use a more efficient index_match_object query to
get the data we need.

This refactors the query to build two nested lists: one of matches (already
matched), and another of selector queries (to be run as a group at the
end). After we've processed all the queries, we gather together all the
results together.